refactor(logitech): move the haptic codec to the package entry point
The wire format belongs in src/logitech/ and reaches consumers through the
Logitech entry point; only feature discovery, WebHID requests and read-back
verification stay in the driver. Following the mode-status module put it
under src/drivers/ instead, which leaves the already-large HID++ driver
owning the packet layout too.

The codec folder uses .js extensions on relative imports and the driver
folder uses .ts, so the moved file and its test follow the folder they now
live in.
