+++ src/logitech/haptics.ts
